+++ lib/rubocop/cop/style/trivial_accessors.rb
@@ -11,12 +11,25 @@ module Rubocop
         def on_def(node)
           method_name, args, body = *node
 
-          kind = if body && body.type == :ivar
+          check(node, method_name, args, body)
+
+          super
+        end
+
+        def on_defs(node)
+          _scope, method_name, args, body = *node
+
+          check(node, method_name, args, body)
+
+          super
+        end
+
+        private
+
+        def check(node, method_name, args, body)
+          kind = if trivial_reader?(method_name, args, body)
                    'reader'
-                 elsif args.children.size == 1 &&
-                       body && body.type == :ivasgn &&
-                       body.children[1] && body.children[1].type == :lvar &&
-                       method_name != :initialize
+                 elsif trivial_writer?(method_name, args, body)
                    'writer'
                  end
           if kind
@@ -24,7 +37,33 @@ module Rubocop
                         sprintf(MSG, kind, kind))
           end
 
-          super
+        end
+
+        def exact_name_match?
+          TrivialAccessors.config['ExactNameMatch']
+        end
+
+        def trivial_reader?(method_name, args, body)
+          return false unless args.children.size == 0
+
+          return false unless body && body.type == :ivar
+
+          exact_name_match? ? names_match?(method_name, body) : true
+        end
+
+        def trivial_writer?(method_name, args, body)
+          return false unless args.children.size == 1 &&
+            body && body.type == :ivasgn &&
+            body.children[1] && body.children[1].type == :lvar &&
+            method_name != :initialize
+
+          exact_name_match? ? names_match?(method_name, body) : true
+        end
+
+        def names_match?(method_name, body)
+          ivar_name, = *body
+
+          method_name.to_s.chomp('=') == ivar_name[1..-1]
         end
       end
     end